fix(dataset_tokenizer): Respect `-uint32_enforce`; only ever promote
This change fixes -uint32_enforce so that it actually produces
output files with 32-bit tokens. Previously, it produced 16-bit files
unless -show_contexts was also specified, plus a few other conditions.

Furthermore, this change switches around the behaviour
of -uint32_enforce.
Previously, the logic for choosing 16-bit or 32-bit output was:
> Always use 16-bit, unless -uint32_enforce was specified,
> AND the encoder is known in the WriteContexts call,
> AND the encoder has a vocab larger than 65536 tokens

Now, the logic is:
> Use 16-bit unless -uint32_enforce was specified, then use 32-bit
> OR if the encoder has a vocab larger than 65536, use 32-bit anyway
> BUT if unspecified and the encoder is unknown, raise an error

It additionally emits a warning when promoting to 32-bit output
without the -uint32_enforce flag, because it has a major impact
on how the file will be read, and the token width is not tagged
in the file itself.

This also fixes the -show_contexts flag, which was broken in that
it didn't show any contexts, as its implementation was mistakenly
removed when -uint32_enforce was added.