fix(dataset_tokenizer): Strip BOS/EOS tokens for `getAndCheckToken`
getAndCheckToken would not recognize a string as a valid token
even if it encoded to a single real token if BOS/EOS markers were tacked
onto the encoded representation. This change checks for that case,
and extracts the real token from the group if found.

This is necessary to be able to set "\n" as a special token while using
the Llama3 tokenizer vocabulary, which treats "\n" as a single token,
but uses an alternative representation for it in its vocab definitions.